+++ spec/spec_helper.rb
@@ -2,5 +2,6 @@ $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 
 require 'rubygems'
 require 'rspec'
+require 'tmpdir'
 require 'fui'